fix: Resolve release/v3.0.0 conflicts and migrate new layers to Keras 3
- Apply hash indexer null behaviour (PR #41): reserve index 0 for nulls,
    num_bins > 1 validation, +1 offset in hash_index, min_hash_index,
bloom_encode
  - Apply allow layer/output names equal (PR #42): remove IdentityLayer
wrapping
    from pipeline_graph, skip self-loops in graph edges
  - Migrate ArrayReduceMax and PairwiseCosineSimilarity as multi-backend
layers
    using keras.ops (PR #45)
  - Fix dtype bug in divide_no_nan (x.dtype -> y.dtype)

+++ src/kamae/keras/core/utils/ops_utils.py
@@ -34,5 +34,5 @@ def divide_no_nan(x: Tensor, y: Tensor) -> Tensor:
     :param y: Denominator tensor
     :returns: Result of x / y, with 0 where y == 0
     """
-    is_zero = ops.equal(y, ops.convert_to_tensor(0.0, dtype=x.dtype))
+    is_zero = ops.equal(y, ops.convert_to_tensor(0.0, dtype=y.dtype))
     return ops.where(is_zero, ops.zeros_like(x), ops.divide(x, y))
